requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-filter==1.0.4
django-tables2==1.7.1
django-widget-tweaks==1.4.1
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2
static3==0.7.0
